requirements.txt
transformers==4.49.0
streamlit==1.37.1
pillow==10.4.0
protobuf==4.25.3
Jinja2==3.1.4
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0